app/requirements.txt
amqp==2.6.1
attrs==20.3.0
backcall==0.2.0
billiard==3.6.3.0
blis==0.4.1
bson==0.5.8
catalogue==1.0.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.4
dataclasses==0.6
decorator==4.4.2
Django==2.2.9
django-celery-results==1.2.1
djangorestframework==3.9.2
djongo==1.3.0
drf-yasg==1.13.0
en-core-sci-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.3.0/en_core_sci_md-0.3.0.tar.gz
flower==0.9.5
humanize==3.1.0
idna==2.10
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
kombu==4.6.11
lxml==4.6.1
MarkupSafe==1.1.1
murmurhash==1.0.4
nmslib==2.0.6
numpy==1.19.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psutil==5.7.3
ptyprocess==0.6.0
pybind11==2.6.0
Pygments==2.7.2
pymongo==3.11.0
pyrsistent==0.17.3
pysbd==0.3.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
scispacy==0.3.0
six==1.15.0
spacy==2.3.2
sqlparse==0.2.4
srsly==1.0.3
swagger-spec-validator==2.7.3
thinc==7.4.1
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
wasabi==0.8.0
wcwidth==0.2.5
